% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_util.R
\name{add_ngrams}
\alias{add_ngrams}
\title{Add ngrams of multiple sizes to given melody data frame}
\usage{
add_ngrams(mel_data, columns, N, override = FALSE)
}
\arguments{
\item{mel_data}{Melody data frame}

\item{N}{(vector of positive integers) N-gram lengths}

\item{override}{(logical) Whether to override existing ngram columns}
}
\value{
melody data frame with added n
}
\description{
Add ngrams of multiple sizes to given melody data frame
}
