% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_sim_measure.R
\name{benchmark_sim_measure}
\alias{benchmark_sim_measure}
\title{Benchmark similarity measures}
\usage{
benchmark_sim_measure(
  sim_measures = c("ngrukkon", "diffed", "rawed"),
  benchmark_corpus = kinder_full,
  limit_corpus_N = NULL,
  return_sim_values = FALSE,
  plot = FALSE
)
}
\arguments{
\item{plot}{}
}
\description{
Benchmark similarity measures
}
\examples{


ex1 <- benchmark_sim_measure(limit_corpus_N = 20L)

ex2 <- benchmark_sim_measure(limit_corpus_N = 20L, return_sim_values = TRUE, plot = TRUE)




}
