% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{benchmark_sim_measure_on_muel_frieler_2004}
\alias{benchmark_sim_measure_on_muel_frieler_2004}
\title{Benchmark against Müllensiefen & Frieler 2004.}
\usage{
benchmark_sim_measure_on_muel_frieler_2004(sim_measure = "opti3")
}
\arguments{
\item{sim_measure}{}
}
\description{
Benchmark against Müllensiefen & Frieler 2004.
}
