% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{fix_mcsv2}
\alias{fix_mcsv2}
\title{MeloSpyGUI exports MCSV2 with too many quotations marks around the strings in the signature column
as a results MeloSpyGUI throws and error when reading in such  MCSV2 files.
Here is a simple function that fixes that.}
\usage{
fix_mcsv2(fnames, outdir)
}
\description{
MeloSpyGUI exports MCSV2 with too many quotations marks around the strings in the signature column
as a results MeloSpyGUI throws and error when reading in such  MCSV2 files.
Here is a simple function that fixes that.
}
