% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setbased_lowlevel.R
\name{get_all_ngrams}
\alias{get_all_ngrams}
\title{Get all ngrams from a given vector}
\usage{
get_all_ngrams(x, N = 3, collapse = ",", keep_length = FALSE)
}
\arguments{
\item{N}{}
}
\description{
Get all ngrams from a given vector
}
