% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{get_implicit_harmonies}
\alias{get_implicit_harmonies}
\title{get harmonies via the Krumhansl-Schmuckler algorithm}
\usage{
get_implicit_harmonies(
  pitch_vec,
  segmentation = NULL,
  weights = NULL,
  only_winner = TRUE,
  fast_algorithm = TRUE
)
}
\arguments{
\item{pitch_vec}{(integer) Pitch vector}

\item{segmentation}{(factor) Segmentation as grouping variable}

\item{only_winner}{(logical) Return only winning harmony}

\item{fast_algorithm}{(logical) Use, fast matrix-base algorithm}
}
\description{
get harmonies via the Krumhansl-Schmuckler algorithm
}
