% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_bank_to_mel_list.R
\name{item_bank_to_mel_list}
\alias{item_bank_to_mel_list}
\title{Convert an itembankr itembank to a list of melsim objects or mutate as new column}
\usage{
item_bank_to_mel_list(item_bank, add_as_column_to_item_bank = FALSE)
}
\arguments{
\item{add_as_column_to_item_bank}{}
}
\description{
Convert an itembankr itembank to a list of melsim objects or mutate as new column
}
