% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melsim.R
\name{melsim}
\alias{melsim}
\title{Compute the melodic similarity two melodies}
\usage{
melsim(
  melody1,
  melody2 = NULL,
  similarity_measures = get_sim_measures(),
  paired = FALSE,
  verbose = T,
  with_progress = TRUE,
  with_checks = TRUE,
  name = "--"
)
}
\arguments{
\item{melody1}{Vector of melody objects or file names - for comparison.}

\item{melody2}{Vector of melody objects or file names - for comparison.}

\item{similarity_algorithm}{Which similarity algorithm to use.}

\item{ngram_length}{For set-based similarity, the N-gram length.}
}
\description{
Compute the melodic similarity two melodies
}
