% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_midi.R
\name{midi_file_to_notes_and_durations}
\alias{midi_file_to_notes_and_durations}
\title{Get notes and durations from a MIDI file}
\usage{
midi_file_to_notes_and_durations(
  midi_file,
  string_df = TRUE,
  produce_extra_melodic_features = TRUE
)
}
\arguments{
\item{produce_extra_melodic_features}{}
}
\description{
Get notes and durations from a MIDI file
}
