% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{motifator}
\alias{motifator}
\title{New segmentation function, does a rather fine segmentation, in fact, it looks for motifs not segments
On the Kinder set, the algorithm receives an median F1 of .65, whereas
itembankr::get_segmentation achvievs a median F1 of .38}
\usage{
motifator(mel_obj, threshold = 3)
}
\description{
New segmentation function, does a rather fine segmentation, in fact, it looks for motifs not segments
On the Kinder set, the algorithm receives an median F1 of .65, whereas
itembankr::get_segmentation achvievs a median F1 of .38
}
