% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_pmi.R
\name{pmi_original}
\alias{pmi_original}
\title{PMI similarity measure}
\usage{
pmi_original(query_pitch, target_pitch, gapOpening = 12, gapExentsion = 6)
}
\arguments{
\item{query_pitch}{(integer vector) query pitch sequence}

\item{target_pitch}{(integer vector) target pitch sequence}
}
\description{
PMI similarity measure
}
