% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_musicxml.R
\name{read_musicxml}
\alias{read_musicxml}
\title{read_musicxml
Simple reader for MusicXML files, works only in the simplest cases and only with most basic information}
\usage{
read_musicxml(musicxml_file)
}
\arguments{
\item{musicxml_file}{(string)}
}
\value{
tibbel with pitch, onset, duration
}
\description{
read_musicxml
Simple reader for MusicXML files, works only in the simplest cases and only with most basic information
}
