% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{sim_measures_from_yaml}
\alias{sim_measures_from_yaml}
\title{sim_measures_from_yaml
Read similarity measur definitions from a YAML and construct a list of SimilaryMeasure objects}
\usage{
sim_measures_from_yaml(fname)
}
\arguments{
\item{fname}{(character) Filename}
}
\value{
list of SimilarityMeasure objects
}
\description{
sim_measures_from_yaml
Read similarity measur definitions from a YAML and construct a list of SimilaryMeasure objects
}
