% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch.R
\name{trans}
\alias{trans}
\title{Transpose a MIDI pitch, or set of MIDI pitches.}
\usage{
trans(midi_pitch, transpose)
}
\arguments{
\item{midi_pitch}{The MIDI pitch or pitches to transpose.}

\item{transpose}{The amount to transpose by as an integer (positive or negative).}
}
\description{
Transpose a MIDI pitch, or set of MIDI pitches.
}
