% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{update_melodies}
\alias{update_melodies}
\title{update_melodies
Updates a list of melody objects to the current version of the melody object (as R6 objects keep their bindings, once created)}
\usage{
update_melodies(mel_list, force = T)
}
\arguments{
\item{mel_list}{list of Melody objects}

\item{force}{(logical) Do it no matter what, otherwise only if version of melsim has changed}
}
\description{
update_melodies
Updates a list of melody objects to the current version of the melody object (as R6 objects keep their bindings, once created)
}
