% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{update_sim_mat}
\alias{update_sim_mat}
\title{update_sim_mat
Updates a SimilarityMatrix object}
\usage{
update_sim_mat(sim_mat, force = T)
}
\arguments{
\item{sim_mat}{SimilarityMatrix object}

\item{force}{(logical) Do it no matter what, otherwise only if version of melsim has changed}
}
\description{
update_sim_mat
Updates a SimilarityMatrix object
}
