% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melsim.R
\name{melsim}
\alias{melsim}
\title{Compute the melodic similarity two melodies}
\usage{
melsim(
  melody1,
  melody2,
  input_type = c("mcsv_file", "named_list", "midi_file", "musicxml_file",
    "monophonic_audio_file", "humdrum_file"),
  similarity_algorithm = c("edit_distance", "set_based", "opti3", "needleman_wunsch",
    "compression_distance_gzip", "correlation", get_proxy_sim_measures()),
  ngram_length = 3L
)
}
\arguments{
\item{melody1}{The first melody - or set of melodies - for comparison.}

\item{melody2}{The second melody - or set of melodies - for comparison.}

\item{input_type}{The format in which the melodies are being represented initially.}

\item{similarity_algorithm}{Which similarity algorithm to use.}

\item{ngram_length}{For set-based similarity, the N-gram length.}
}
\description{
Compute the melodic similarity two melodies
}
